Change 'j' hotkey to always go to next article.
Currently if you use Space to scroll an entry you get stuck in a mode
where j/k only scroll the entry up and down. Previously there were two
ways to get focus back on the EntryList: repeatedly hit space until you
get to the end of the article, change focus back to the Entry list with
'h' or left arrow. This patch overrides the 'j' hotkey scrolling
behavior and forces it to function globally as NextUnreadEntry.